Move tests for user's authentication functionality to the UserSpec class
